move Menu from internal to public widget package
