+++ widget/menu_desktop_test.go
@@ -1,19 +1,21 @@
-// +build mobile
+// +build !mobile
 
 package widget_test
 
 import (
+	"image/color"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
-
 	"fyne.io/fyne"
-	"fyne.io/fyne/internal/widget"
+	"fyne.io/fyne/internal/cache"
 	"fyne.io/fyne/test"
 	"fyne.io/fyne/theme"
+	"fyne.io/fyne/widget"
+
+	"github.com/stretchr/testify/assert"
 )
 
-func TestMenu_ItemWithChildTapped(t *testing.T) {
+func TestMenu_ItemHovered(t *testing.T) {
 	sub1 := fyne.NewMenuItem("sub1", nil)
 	sub1.ChildMenu = fyne.NewMenu("",
 		fyne.NewMenuItem("sub1 A", nil),
@@ -42,30 +44,59 @@ func TestMenu_ItemWithChildTapped(t *testing.T) {
 	size := m.MinSize()
 	m.Resize(size)
 
+	mi := m.Items[0].(*widget.MenuItem)
+	r1 := cache.Renderer(mi)
+	assert.Equal(t, color.Transparent, r1.BackgroundColor())
+	mi.MouseIn(nil)
+	assert.Equal(t, theme.HoverColor(), r1.BackgroundColor())
+	mi.MouseOut()
+	assert.Equal(t, color.Transparent, r1.BackgroundColor())
+
 	sub1Widget := m.Items[3].(*widget.MenuItem)
 	assert.Equal(t, sub1, sub1Widget.Item)
 	sub2Widget := m.Items[4].(*widget.MenuItem)
 	assert.Equal(t, sub2, sub2Widget.Item)
 	assert.False(t, sub1Widget.Child.Visible(), "submenu is invisible initially")
 	assert.False(t, sub2Widget.Child.Visible(), "submenu is invisible initially")
-	test.Tap(sub1Widget)
-	assert.True(t, sub1Widget.Child.Visible(), "tapping item shows submenu")
+	sub1Widget.MouseIn(nil)
+	assert.True(t, sub1Widget.Child.Visible(), "hovering item shows submenu")
 	assert.False(t, sub2Widget.Child.Visible(), "other Child menu stays hidden")
-	test.Tap(sub2Widget)
-	assert.False(t, sub1Widget.Child.Visible(), "tapping other item hides current submenu")
+	sub1Widget.MouseOut()
+	assert.True(t, sub1Widget.Child.Visible(), "hover out does not hide submenu")
+	assert.False(t, sub2Widget.Child.Visible(), "other Child menu still hidden")
+	sub2Widget.MouseIn(nil)
+	assert.False(t, sub1Widget.Child.Visible(), "hovering other item hides current submenu")
 	assert.True(t, sub2Widget.Child.Visible(), "other Child menu shows up")
 
 	sub2subWidget := sub2Widget.Child.Items[2].(*widget.MenuItem)
 	assert.Equal(t, sub2sub, sub2subWidget.Item)
 	assert.False(t, sub2subWidget.Child.Visible(), "2nd level submenu is invisible initially")
-	test.Tap(sub2subWidget)
+	sub2Widget.MouseOut()
+	sub2subWidget.MouseIn(nil)
 	assert.True(t, sub2Widget.Child.Visible(), "1st level submenu stays visible")
 	assert.True(t, sub2subWidget.Child.Visible(), "2nd level submenu shows up")
+	sub2subWidget.MouseOut()
+	assert.True(t, sub2Widget.Child.Visible(), "1st level submenu still visible")
+	assert.True(t, sub2subWidget.Child.Visible(), "2nd level submenu still visible")
 
-	test.Tap(sub1Widget)
+	sub1Widget.MouseIn(nil)
 	assert.False(t, sub2Widget.Child.Visible(), "1st level submenu is hidden by other submenu")
-	test.Tap(sub2Widget)
+	sub2Widget.MouseIn(nil)
 	assert.False(t, sub2subWidget.Child.Visible(), "2nd level submenu is hidden when re-entering its parent")
+}
+
+func TestMenu_ItemWithChildTapped(t *testing.T) {
+	sub := fyne.NewMenuItem("sub1", nil)
+	sub.ChildMenu = fyne.NewMenu("", fyne.NewMenuItem("sub A", nil))
+	m := widget.NewMenu(fyne.NewMenu("", sub))
+	size := m.MinSize()
+	m.Resize(size)
+
+	subWidget := m.Items[0].(*widget.MenuItem)
+	assert.Equal(t, sub, subWidget.Item)
+	assert.False(t, subWidget.Child.Visible(), "submenu is invisible")
+	test.Tap(subWidget)
+	assert.False(t, subWidget.Child.Visible(), "submenu does not show up")
 }
 
 func TestMenu_Layout(t *testing.T) {
@@ -92,11 +123,11 @@ func TestMenu_Layout(t *testing.T) {
 	menu := fyne.NewMenu("", item1, sep, item2, item3)
 
 	for name, tt := range map[string]struct {
-		windowSize   fyne.Size
-		menuPos      fyne.Position
-		tapPositions []fyne.Position
-		useTestTheme bool
-		wantImage    string
+		windowSize     fyne.Size
+		menuPos        fyne.Position
+		mousePositions []fyne.Position
+		useTestTheme   bool
+		wantImage      string
 	}{
 		"normal": {
 			windowSize: fyne.NewSize(500, 300),
@@ -106,49 +137,49 @@ func TestMenu_Layout(t *testing.T) {
 		"normal with submenus": {
 			windowSize: fyne.NewSize(500, 300),
 			menuPos:    fyne.NewPos(10, 10),
-			tapPositions: []fyne.Position{
+			mousePositions: []fyne.Position{
 				fyne.NewPos(30, 100),
 				fyne.NewPos(100, 170),
 			},
-			wantImage: "menu_layout_mobile_normal_with_submenus.png",
+			wantImage: "menu_layout_normal_with_submenus.png",
 		},
 		"background of active submenu parents resets if sibling is hovered": {
 			windowSize: fyne.NewSize(500, 300),
 			menuPos:    fyne.NewPos(10, 10),
-			tapPositions: []fyne.Position{
+			mousePositions: []fyne.Position{
 				fyne.NewPos(30, 100),  // open submenu
 				fyne.NewPos(100, 170), // open subsubmenu
 				fyne.NewPos(300, 170), // hover subsubmenu item
 				fyne.NewPos(30, 60),   // hover sibling of submenu parent
 			},
-			wantImage: "menu_layout_mobile_background_reset.png",
+			wantImage: "menu_layout_background_reset.png",
 		},
 		"no space on right side for submenu": {
 			windowSize: fyne.NewSize(500, 300),
 			menuPos:    fyne.NewPos(410, 10),
-			tapPositions: []fyne.Position{
+			mousePositions: []fyne.Position{
 				fyne.NewPos(430, 100), // open submenu
 				fyne.NewPos(300, 170), // open subsubmenu
 			},
-			wantImage: "menu_layout_mobile_no_space_on_right.png",
+			wantImage: "menu_layout_no_space_on_right.png",
 		},
 		"no space on left & right side for submenu": {
 			windowSize: fyne.NewSize(200, 300),
 			menuPos:    fyne.NewPos(10, 10),
-			tapPositions: []fyne.Position{
+			mousePositions: []fyne.Position{
 				fyne.NewPos(30, 100),  // open submenu
 				fyne.NewPos(100, 170), // open subsubmenu
 			},
-			wantImage: "menu_layout_mobile_no_space_on_both_sides.png",
+			wantImage: "menu_layout_no_space_on_both_sides.png",
 		},
 		"window too short for submenu": {
 			windowSize: fyne.NewSize(500, 150),
 			menuPos:    fyne.NewPos(10, 10),
-			tapPositions: []fyne.Position{
+			mousePositions: []fyne.Position{
 				fyne.NewPos(30, 100),  // open submenu
 				fyne.NewPos(100, 130), // open subsubmenu
 			},
-			wantImage: "menu_layout_mobile_window_too_short.png",
+			wantImage: "menu_layout_window_too_short.png",
 		},
 		"theme change": {
 			windowSize:   fyne.NewSize(500, 300),
@@ -164,8 +195,8 @@ func TestMenu_Layout(t *testing.T) {
 			w.Resize(tt.windowSize) // SetContent changes window’s size
 			m.Resize(m.MinSize())
 			m.Move(tt.menuPos)
-			for _, pos := range tt.tapPositions {
-				test.TapCanvas(c, pos)
+			for _, pos := range tt.mousePositions {
+				test.MoveMouse(c, pos)
 			}
 			if tt.useTestTheme {
 				test.WithTestTheme(t, func() {
