+++ widget/testdata/menu_layout_background_reset.png
