+++ widget/testdata/menu_layout_mobile_background_reset.png
