+++ widget/testdata/menu_layout_mobile_no_space_on_right.png
