+++ widget/testdata/menu_layout_mobile_window_too_short.png
