+++ widget/testdata/menu_layout_no_space_on_right.png
