+++ widget/testdata/menu_layout_normal.png
