+++ widget/testdata/menu_layout_normal_with_submenus.png
