+++ widget/testdata/menu_layout_theme_changed.png
