+++ widget/testdata/menu_layout_window_too_short.png
